MAINT: remove kolmogorov from functions.json
